+++ src/main/java/com/hetun/datacenter/tools/DateUtils.java
@@ -2,7 +2,7 @@ package com.hetun.datacenter.tools;
 
 public class DateUtils {
 
-    public static Long now() {
-        return System.currentTimeMillis() / 1000;
+    public static Integer now() {
+        return Math.toIntExact(System.currentTimeMillis() / 1000);
     }
 }
